refactor(translation): scaffold and audit split along their own section markers
Both files already narrated their phases in comment rules; each phase is
now a file. Scaffold: differ (signature matching over the full path
set), renderer (edge → parse-refusing .bluebook text), writer
(regenerate in place by shape pair). Audit: layer_one (the bluebook's
own rules), layer_two (the edge against the reference transform),
unfed_report, approval_digest (the human gate's binding). The outside
files keep the structs, the verdict, and check's assembly.
